Refactor workflow architecture.

+++ bin/fim.py
@@ -4,40 +4,53 @@ import pickle
 import time
 from argparse import ArgumentParser
 
+from collections import defaultdict
 from itertools import combinations
 from pathlib import Path
 
 import numpy as np
 from rdflib import Graph, Namespace, Literal
-from rdflib.namespace import RDF, NamespaceManager
+from rdflib.namespace import RDF, RDFS, XSD, DC, NamespaceManager
 from rdflib.term import Node
 
 PROJECT_ROOT = Path(os.environ["HOME"], "code", "Thesis")
 
-TWEETS = Path(PROJECT_ROOT, "tweet-extraction", "first_10M_lines.n3")
+TWEET_EXTRACTION = Path("PROJECT_ROOT", "tweet-extraction")
+TWEETS = Path(TWEET_EXTRACTION, "first_10M_lines.n3")
+TWEETS_WITH_RELATIONS = Path(TWEET_EXTRACTION, "related")
 
 FIM = Path(PROJECT_ROOT, "fim")
-GENERATED = Path(FIM, "generated")
+GRAPHS = Path(FIM, "generated")
 
 UBY_DIR = Path(PROJECT_ROOT, "UBY")
 UBY_NEIGHBORS = Path(UBY_DIR, "uby-neighbors.nt")
 
-DBO = Namespace("http://dbpedia.org/ontology/")
-SCHEMA = Namespace("http://schema.org/")
 NEE = Namespace("http://www.ics.forth.gr/isl/oae/core#")
+SCHEMA = Namespace("http://schema.org/")
 SIOC = Namespace("http://rdfs.org/sioc/ns#")
+SIOC_T = Namespace("http://rdfs.org/sioc/types#")
+ONYX = Namespace("http://www.gsi.dit.upm.es/ontologies/onyx/ns#")
+WNA = Namespace("http://www.gsi.dit.upm.es/ontologies/wnaffect/ns#")
+
 LEMON = Namespace("http://www.monnet-project.eu/lemon#")
+DBO = Namespace("http://dbpedia.org/ontology/")
 
 g = Graph()
 
 
 def load_graphs():
     merge_graph = Graph()
-    for prefix in [("lemon", LEMON),
+    for prefix in [("rdf", RDF),
+                   ("rdfs", RDFS),
+                   ("xsd", XSD),
+                   ("dc", DC),
+                   ("lemon", LEMON),
                    ("nee", NEE),
                    ("dbo", DBO),
-                   ("rdf", RDF),
                    ("schema", SCHEMA),
+                   ("wna", WNA),
+                   ("SIOC_T", SIOC_T),
+                   ("onyx", ONYX),
                    ("sioc", SIOC)]:
         merge_graph.bind(*prefix)
 
@@ -69,7 +82,18 @@ def bag_of_terms(tweet_graph):
     bag = list()
     for triple in tweet_graph.triples((None, None, None)):
         bag.append(triple[0])
+        bag.append(triple[1])
         bag.append(triple[2])
+        bag.append(f"{triple[0]}_{triple[1]}_{triple[2]}")
+
+        # BA :spouse :MA
+        # bag { :BA, :spouse, :MA, ?v0-:spouse-:MA, .... }
+        # the indices of the variable are progressive naturals
+
+
+        bag.append(f"{triple[0]}_{triple[1]}_X")
+        bag.append(f"X_{triple[1]}_{triple[2]}")
+
     return bag
 
 
@@ -92,7 +116,7 @@ def get_tweets(paths):
         if uby_triples.is_file():
             tweet_graph.parse(location=str(uby_triples), format="ttl")
 
-        tweet_graph.serialize(destination=str(Path(GENERATED, f"t{tweet_id}.ttl")), format="turtle")
+        tweet_graph.serialize(destination=str(Path(GRAPHS, f"t{tweet_id}.ttl")), format="turtle")
 
         yield tweet_graph
 
@@ -139,7 +163,7 @@ def column_idx(column, columns):
 
 
 def get_large_1_sets(rows, min_sup=1e-3):
-    supports_selected = {}
+    supports_selected = defaultdict(float)
     # Sum columns and divide by number of transactions
     supports = rows.sum(axis=0) / float(rows.shape[0])
     sets = []
@@ -185,6 +209,14 @@ def select_rows(candidate, rows):
     return rows[selected, :]
 
 
+def compute_support(candidate, rows):
+    # Get rows which contain the candidate set
+    selected = select_rows(candidate, rows)
+
+    # Compute support of the candidate set, i.e. the frequency
+    return selected.shape[0] / float(rows.shape[0])
+
+
 def item_to_string(item):
     return "{" + ', '.join(map(str, item)) + "}"
 
@@ -207,11 +239,8 @@ def frequent_itemsets(rows, min_sup=0.027, max_length=5):
         # Select candidates
         for i, candidate in enumerate(ck):
             print(f"Computing {item_to_string(candidate)} support...\t[{i + 1}/{ck_len}]")
-            # Get rows which contain the candidate set
-            selected = select_rows(candidate, rows)
 
-            # Compute support of the candidate set
-            support = selected.shape[0] / float(rows.shape[0])
+            support = compute_support(candidate, rows)
             supports[candidate] = support
 
             if np.greater_equal(support, min_sup):
@@ -234,6 +263,118 @@ def frequent_itemsets(rows, min_sup=0.027, max_length=5):
     return items, supports
 
 
+# Re-examination of interestingness measures
+# in pattern mining: a unified framework
+# Tianyi Wu · Yuguo Chen · Jiawei Han
+def conditional_probability(supports, a, b):
+    """
+    Convert the support counts into conditional
+    probabilities using the following equations:
+
+    P(a|b) = sup(ab) / sup(b) = sup(ab) / ( sup(ab) + sup(not(a)b) )
+
+    """
+    a_and_b = a & b
+    b_and_not_a = b - a_and_b
+    assert b == b_and_not_a
+    sup_ab = supports[a | b]
+    sup_b = supports[b]
+    if sup_ab == 0.0 or sup_b == 0.0:
+        return 0.0
+    else:
+        return sup_ab / sup_b
+
+
+def all_conf(supports, a, b):
+    # P(a|b)
+    p_a_b = conditional_probability(supports, a, b)
+    # P(b|a)
+    p_b_a = conditional_probability(supports, b, a)
+    return min(p_a_b, p_b_a)
+
+
+def coherence(supports, a, b):
+    # P(a|b)
+    p_a_b = conditional_probability(supports, a, b)
+    # P(b|a)
+    p_b_a = conditional_probability(supports, b, a)
+    return 1.0 / ((1.0 / p_a_b) + (1.0 / p_b_a) - 1)
+
+
+def cosine(supports, a, b):
+    # P(a|b)
+    p_a_b = conditional_probability(supports, a, b)
+    # P(b|a)
+    p_b_a = conditional_probability(supports, b, a)
+    return np.sqrt(p_a_b * p_b_a)
+
+
+def kulc(supports, a, b):
+    # P(a|b)
+    p_a_b = conditional_probability(supports, a, b)
+    # P(b|a)
+    p_b_a = conditional_probability(supports, b, a)
+    return (p_a_b + p_b_a) / 2.0
+
+
+def max_conf(supports, a, b):
+    # P(a|b)
+    p_a_b = conditional_probability(supports, a, b)
+    # P(b|a)
+    p_b_a = conditional_probability(supports, b, a)
+    return max(p_a_b, p_b_a)
+
+
+# A transaction that does not contain an item i is called a null transaction with respect
+# to (w.r.t.) item i.
+
+# If a measure that assesses the association among a set of events
+# being analyzed is affected by the number of transactions that contain none of them
+# (i.e., null-transactions), such a measure is unlikely to be so interesting.
+
+# Thus it is desirable to select a measure that is not influenced by
+# the number of null transactions, i.e., null-invariant.
+
+def compute_metrics(supports, freq_set, subsets, min_confidence):
+    """
+    create the rules and returns the rules info and the rules's
+    right hand side (used for generating the next round of rules)
+    if it surpasses the minimum confidence threshold
+    """
+    rules = []
+    right_hand_side = []
+
+    # we can constrain the rule gen search
+    # {...} = > { ?v0 - dbo: spouse -?v1}
+
+    # we can add prefixes to rules contents to figure out if a rule matches a pattern
+    # bag = {term_: BA
+    # tp =?v0 - dbo: spouse -?v1
+    # {} = > {tp}
+
+    for rhs in subsets:
+        # create the left hand side of the rule
+        # and add the rules if it's greater than
+        # the confidence threshold
+        lhs = freq_set - rhs
+        conf = supports[freq_set] / supports[lhs]
+        # if conf >= min_confidence:
+        lift = conf / supports[rhs]
+        rules_info = (lhs,
+                      rhs,
+                      conf,
+                      lift,
+                      all_conf(supports, lhs, rhs),
+                      coherence(supports, lhs, rhs),
+                      cosine(supports, lhs, rhs),
+                      kulc(supports, lhs, rhs),
+                      max_conf(supports, lhs, rhs))
+        rules.append(rules_info)
+        right_hand_side.append(rhs)
+
+    return rules, right_hand_side
+
+
 def create_rules(freq_items, supports, min_confidence):
     """
     create the association rules, the rules will be a list.
@@ -254,7 +395,7 @@ def create_rules(freq_items, supports, min_confidence):
             # start with creating rules for single item on
             # the right hand side
             subsets = [frozenset([item]) for item in freq_set]
-            rules, right_hand_side = compute_conf(supports, freq_set, subsets, min_confidence)
+            rules, right_hand_side = compute_metrics(supports, freq_set, subsets, min_confidence)
             association_rules.extend(rules)
 
             # starting from 3-itemset, loop through each length item
@@ -268,81 +409,53 @@ def create_rules(freq_items, supports, min_confidence):
                 k = 1
                 while len(right_hand_side[0]) < len(freq_set) - 1:
                     ck = candidate_itemsets(right_hand_side, k)
-                    rules, right_hand_side = compute_conf(supports, freq_set, ck, min_confidence)
+                    rules, right_hand_side = compute_metrics(supports, freq_set, ck, min_confidence)
                     association_rules.extend(rules)
                     k += 1
 
     return association_rules
 
 
-def compute_conf(supports, freq_set, subsets, min_confidence):
-    """
-    create the rules and returns the rules info and the rules's
-    right hand side (used for generating the next round of rules)
-    if it surpasses the minimum confidence threshold
-    """
-    rules = []
-    right_hand_side = []
-
-    for rhs in subsets:
-        # create the left hand side of the rule
-        # and add the rules if it's greater than
-        # the confidence threshold
-        lhs = freq_set - rhs
-        conf = supports[freq_set] / supports[lhs]
-        if conf >= min_confidence:
-            lift = conf / supports[rhs]
-            rules_info = lhs, rhs, conf, lift
-            rules.append(rules_info)
-            right_hand_side.append(rhs)
-
-    return rules, right_hand_side
-
-
 def main():
-    # get itemsets
-    tweets = get_tweets(Path(TWEETS).iterdir())
-    bags = map(bag_of_terms, tweets)
+    rows_npy_path = Path(FIM, "rows.npy")
+    last_l_pickle = Path(FIM, "last.pickle")
+    supports_pickle = Path(FIM, "supports.pickle")
+    rules_pickle_path = Path(FIM, "rules.pickle")
 
-    columns, rows = build_db(bags)
+    if not rows_npy_path.is_file():
+        tweets = get_tweets(Path(TWEETS_WITH_RELATIONS).iterdir())
+        bags = map(bag_of_terms, tweets)
+        columns, rows = build_db(bags)
 
-    np.save(Path(FIM, "rows"), rows)
-    np.save(Path(FIM, "columns"), columns, allow_pickle=True)
+        np.save(Path(FIM, "rows"), rows)
+        np.save(Path(FIM, "columns"), columns, allow_pickle=True)
 
-    columns = np.load(Path(FIM, "columns.npy"), allow_pickle=True)
-    rows = np.load(Path(FIM, "rows.npy"))
+    if (not last_l_pickle.is_file()) or (not supports_pickle.is_file()):
+        rows = np.load(rows_npy_path)
 
-    last_l, supports = frequent_itemsets(rows)
+        last_l, supports = frequent_itemsets(rows)
 
-    with open(Path(FIM, "last.pickle"), "wb") as fp:
-        pickle.dump(last_l, fp)
+        with open(last_l_pickle, "wb") as fp:
+            pickle.dump(last_l, fp)
 
-    with open(Path(FIM, "supports.pickle"), "wb") as fp:
-        pickle.dump(supports, fp)
+        with open(supports_pickle, "wb") as fp:
+            pickle.dump(supports, fp)
 
-    with open(Path(FIM, "last.pickle"), "rb") as fp:
-        last_l = pickle.load(fp)
-    with open(Path(FIM, "supports.pickle"), "rb") as fp:
-        supports = pickle.load(fp)
+    if not rules_pickle_path.is_file():
+        with open(last_l_pickle, "rb") as fp:
+            last_l = pickle.load(fp)
+        with open(supports_pickle, "rb") as fp:
+            supports = pickle.load(fp)
 
-    rules = create_rules(last_l, supports, min_confidence=0.01)
-    with open(Path(FIM, "rules.pickle"), "wb") as fp:
-        pickle.dump(rules, fp)
-
-    join_urls = lambda urls: "\n".join(columns[i] for i in urls)
-    for lhs, rhs, conf, lift in sorted(rules, key=lambda info: info[3]):
-        print("---------------------------------------------------------")
-        left = join_urls(lhs)
-        right = join_urls(rhs)
-        msg = f"{left}\n\n=>\n\n{right}\n\nwith confidence {conf} and lift {lift}"
-        print(msg)
-        print("---------------------------------------------------------")
+        rules = create_rules(last_l, supports, min_confidence=0.01)
+        with open(rules_pickle_path, "wb") as fp:
+            pickle.dump(rules, fp)
 
 
 if __name__ == "__main__":
     parser = ArgumentParser()
     parser.add_argument("-t", "--tweets-graph", type=Path, required=True,
-                        help="Path of .n3 graph representing the tweets and entities.")
+                        help="Path of the TweetsKB graph representing the tweets and entities.")
     parser.add_argument("-r", "--related", type=Path, required=True,
                         help="Path of directory the graphs of the tweets that have at least one relationship between "
                              "their entities.")
@@ -354,11 +467,12 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
     FIM = args.out_dir
-    GENERATED = Path(FIM, "generated")
+    GRAPHS = Path(FIM, "graphs")
     UBY_NEIGHBORS = args.uby_neighbors
     TWEETS = args.tweets_graph
+    TWEETS_WITH_RELATIONS = args.related
 
-    GENERATED.mkdir(exist_ok=True, parents=True)
+    print(f"Output will be generated at {FIM}")
 
     g = load_graphs()
